throw exception when formatted value is automatic
When the formatted value is set to "automatic" then this module throws an error. The exception gives a hint what is wrong.
It would be even be better to get the field with a different method if the image field returns a string, but I only wanted a quick fix.